freed up a ton of SRAM memory by fixing the bug that made us need to store 8-bit data in 16-bit arrays for communication with the external memory. We now have 11200 bytes back that we we wasting (out of 64000 total SRAM bytes available)

+++ MOUSE_DUAL/MOUSE_DUAL/src/memory_spi.h
@@ -31,8 +31,8 @@ extern unsigned char startupStateLoadPending;
 // Memory SPI
 
 int memorySPICheckIfBusy(void);
-void memorySPIWrite(int sector, int page, uint16_t* buffer, int numBytes);
-void memorySPIRead(int sector, int page, uint16_t* buffer, int numBytes);
+void memorySPIWrite(int sector, int page, uint8_t* buffer, int numBytes);
+void memorySPIRead(int sector, int page, uint8_t* buffer, int numBytes);
 void memorySPIEraseSector(uint16_t sector); // Sector can be 0-2047
 void memorySPIEraseBlock(uint16_t block); // Block 0 - 127
